move make_matrix() into utils.py
