+++ find_duplicates_test.py
@@ -3,6 +3,7 @@ import numpy
 import unittest
 
 import find_duplicates
+import utils
 import visual_diff
 
 
@@ -10,7 +11,7 @@ class TestGetLengths(unittest.TestCase):
     def setUp(self):
         contents = 'print(hello(1, 2)) and print(hello("hi", 2))'
         data = visual_diff.get_tokens(contents, "python")
-        self.matrix = visual_diff.make_matrix(data.tokens, data.tokens)
+        self.matrix = utils.make_matrix(data.tokens, data.tokens)
 
     @staticmethod
     def debug_differences(expected, actual):
